major refactoring of Code

+++ Language/Brainfuck/Command.hs
@@ -2,7 +2,7 @@ module Language.Brainfuck.Command (Command (..)) where
 
 import Data.Int
 
-data Command = IncrementPointer Int16
+data Command = IncrementPointer Int32
              | IncrementByte Int32
              | OutputByte
              | InputByte
